[BEAM-8137] Refactor External Worker Service to be part of sdks-java-harness.
Move external worker service from runners/portability/java to sdks/java/harness
Move external worker service depdended classes from runners/java-fn-execution to sdks/java/fn-execution

In order that external worker service can be launched in worker container's boot script where only beam-sdks-java-harness is intially included in the image.
