+++ runners/java-fn-execution/src/main/java/org/apache/beam/runners/fnexecution/state/GrpcStateService.java
@@ -27,7 +27,7 @@ import java.util.concurrent.ConcurrentMap;
 import org.apache.beam.model.fnexecution.v1.BeamFnApi.StateRequest;
 import org.apache.beam.model.fnexecution.v1.BeamFnApi.StateResponse;
 import org.apache.beam.model.fnexecution.v1.BeamFnStateGrpc;
-import org.apache.beam.runners.fnexecution.FnService;
+import org.apache.beam.sdk.fn.FnService;
 import org.apache.beam.vendor.grpc.v1p36p0.io.grpc.stub.ServerCallStreamObserver;
 import org.apache.beam.vendor.grpc.v1p36p0.io.grpc.stub.StreamObserver;
