+++ runners/java-fn-execution/src/test/java/org/apache/beam/runners/fnexecution/control/FnApiControlClientPoolServiceTest.java
@@ -32,9 +32,9 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import org.apache.beam.model.fnexecution.v1.BeamFnApi;
 import org.apache.beam.model.fnexecution.v1.BeamFnApi.InstructionRequest;
 import org.apache.beam.model.fnexecution.v1.BeamFnControlGrpc;
-import org.apache.beam.runners.fnexecution.GrpcContextHeaderAccessorProvider;
-import org.apache.beam.runners.fnexecution.GrpcFnServer;
-import org.apache.beam.runners.fnexecution.InProcessServerFactory;
+import org.apache.beam.sdk.fn.GrpcContextHeaderAccessorProvider;
+import org.apache.beam.sdk.fn.GrpcFnServer;
+import org.apache.beam.sdk.fn.InProcessServerFactory;
 import org.apache.beam.sdk.util.MoreFutures;
 import org.apache.beam.vendor.grpc.v1p36p0.io.grpc.Status;
 import org.apache.beam.vendor.grpc.v1p36p0.io.grpc.StatusException;
