+++ runners/java-fn-execution/src/test/java/org/apache/beam/runners/fnexecution/provisioning/StaticGrpcProvisionServiceTest.java
@@ -25,9 +25,9 @@ import org.apache.beam.model.fnexecution.v1.ProvisionApi.GetProvisionInfoRespons
 import org.apache.beam.model.fnexecution.v1.ProvisionApi.ProvisionInfo;
 import org.apache.beam.model.fnexecution.v1.ProvisionServiceGrpc;
 import org.apache.beam.model.fnexecution.v1.ProvisionServiceGrpc.ProvisionServiceBlockingStub;
-import org.apache.beam.runners.fnexecution.GrpcContextHeaderAccessorProvider;
-import org.apache.beam.runners.fnexecution.GrpcFnServer;
-import org.apache.beam.runners.fnexecution.InProcessServerFactory;
+import org.apache.beam.sdk.fn.GrpcContextHeaderAccessorProvider;
+import org.apache.beam.sdk.fn.GrpcFnServer;
+import org.apache.beam.sdk.fn.InProcessServerFactory;
 import org.apache.beam.vendor.grpc.v1p36p0.com.google.protobuf.ListValue;
 import org.apache.beam.vendor.grpc.v1p36p0.com.google.protobuf.NullValue;
 import org.apache.beam.vendor.grpc.v1p36p0.com.google.protobuf.Struct;
