+++ runners/samza/src/main/java/org/apache/beam/runners/samza/SamzaJobServerDriver.java
@@ -17,7 +17,7 @@
  */
 package org.apache.beam.runners.samza;
 
-import org.apache.beam.runners.fnexecution.ServerFactory;
+import org.apache.beam.sdk.fn.ServerFactory;
 import org.apache.beam.runners.jobsubmission.JobServerDriver;
 import org.apache.beam.sdk.io.FileSystems;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
