+++ sdks/java/fn-execution/src/main/java/org/apache/beam/sdk/fn/InProcessServerFactory.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.runners.fnexecution;
+package org.apache.beam.sdk.fn;
 
 import java.io.IOException;
 import java.util.List;
